front/requirements.txt
azure-functions==1.15.0
streamlit==1.22.0
openai==0.27.2
matplotlib==3.6.3
plotly==5.12.0
scipy==1.10.0
scikit-learn==1.2.0
redis==4.5.3
transformers==4.25.1
python-dotenv==0.21.0
azure-ai-formrecognizer==3.3.0b1
azure-storage-blob==12.14.1
requests==2.28.2
tiktoken==0.2.0
azure-storage-queue==12.5.0
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
sqlmodel==0.0.8
pymssql==2.2.7
langchain==0.0.232
beautifulsoup4==4.12.0
sqlalchemy==1.4.31
streamlit==1.22.0